Fix issue with Date processing: 'date format was not recognized'
Related issue: https://github.com/tritum/repeatable_form_elements/issues/4

I have not found a way to set this ProcessingRules directly via form extension API. That why two variants are included (one active, one commented out).
The activated variant does not require modification in the form extension but uses, because of permission restriction on API methods, reflection. It is only a work around until a way is given to set ProcessingRules via API method.

The second variant (commented out) with direct method call is to prefer. One has to switch to this variant if the new method in the form extension is available.