Remove siblings of file input
If a file upload field has a session value, the UploadResourceViewHelper used in the FileUpload and ImageUpload partials creates a hidden field containing the newly created file reference. In addition a container with a preview of the uploaded file is created. If you clone the file input now these additional elements have to be removed...